Add files via upload
#include<iostream>:-This line includes the standard input/output library in the program. It allows you to use input and output functions like cin and cout.
int main():-This is the main function where the program execution begins.
The purpose of this program is to act as a simple calculator. 
It performs basic arithmetic operations like addition, subtraction, multiplication, and division.
When you run this program, it will ask you to enter two integers.
After you enter the two integers, it will perform the operations and display the result.