Add files via upload
#include<iostream>:-This line includes the standard input/output library in the program. It allows you to use input and output functions like cin and cout.
using namespace std means that we can use names for objects and variables from the standard library.
int main():-This is the main function where the program execution begins.
This  program checks if a number entered by the user is between 10 and 20. 
If the number is between 10 and 20, it prints: “The number is between 10 and 20.”
Otherwise, it prints: “The number is not between 10 and 20.”
This program is a simple example to check a range of values.