Add files via upload
#include<iostream>:-This line includes the standard input/output library in the program. It allows you to use input and output functions like cin and cout.
using namespace std means that we can use names for objects and variables from the standard library.
int main():-This is the main function where the program execution begins.
This program is a simple calculator that performs basic arithmetic operations (addition, subtraction, multiplication, and division) based on the operator entered by the user. 
The switch statement in this program checks the operator entered by the user and performs the corresponding arithmetic operation (addition, subtraction, multiplication, or division). If the operator is not recognized, it prints an error message.