Add files via upload
#include<iostream>:-This line includes the standard input/output library in the program. It allows you to use input and output functions like cin and cout.
using namespace std means that we can use names for objects and variables from the standard library.
int main():-This is the main function where the program execution begins.
This program demonstrates various string manipulation operations.
Concatenation, Substring, Find, Replace, Insert, Erase, To Uppercase, To Lowercase.
Concatenation: This operation combines two strings into one.
Substring: This extracts a portion of a string.
Find: This locates the position of a substring within a string.
Replace: This replaces a part of the string with another string. 
Insert: This inserts a substring at a specified position in a string. 
Erase: This removes a portion of the string. 
To Uppercase: This converts all characters in a string to uppercase. 
To Lowercase: This converts all characters in a string to lowercase.