Add files via upload
#include<iostream>:-This line includes the standard input/output library in the program. It allows you to use input and output functions like cin and cout.
using namespace std means that we can use names for objects and variables from the standard library.
int main():-This is the main function where the program execution begins.
This program calculates the factorial of a non-negative integer using both iterative and recursive methods. 
This program demonstrates two different methods to calculate the factorial of a number: iteratively using a loop and recursively using function calls.
It also handles invalid input by checking if the entered number is negative.