Add files via upload
#include<iostream>:-This line includes the standard input/output library in the program. It allows you to use input and output functions like cin and cout.
using namespace std means that we can use names for objects and variables from the standard library.
int main():-This is the main function where the program execution begins.
Three integer variables a, b, and c are declared to store the user’s input.
The first if statement checks if a is greater than or equal to both b and c. If true, it prints that a is the largest number.
The second if statement checks if b is greater than or equal to both a and c. If true, it prints that b is the largest number.
The third if statement checks if c is greater than or equal to both a and b. If true, it prints that c is the largest number.