Add files via upload
#include<iostream>:-This line includes the standard input/output library in the program. It allows you to use input and output functions like cin and cout.
using namespace std means that we can use names for objects and variables from the standard library.
int main():-This is the main function where the program execution begins.
divisor, dividend, quotient, and remainder are declared as integer variables. 
The program ask the user to enter the dividend and divisor values using.
The quotient is calculated by dividing the dividend by the divisor using the / operator.
The remainder is calculated using the modulus operator %, which gives the remainder of the division.
The program outputs the quotient and remainder.